requirements.txt
# Selenium and WebDriver
selenium==4.15.2

# Data handling
pandas==2.1.4
openpyxl==3.1.2

# Progress bars
tqdm==4.66.1

# Environment variables
python-dotenv==1.0.0

# Automation
pyautogui==0.9.54

# Network utilities
requests==2.31.0

# Additional utilities
Pillow==10.1.0